ENH: delay expensive imports
